webpack config wip
