+++ package.json
@@ -5,7 +5,9 @@
   "main": "index.js",
   "scripts": {
     "prestart": "babel-node buildScripts/startMessage.js",
-    "start": "npm-run-all --parallel open:src lint:watch test:watch start-mockapi",
+    "start": "npm-run-all --parallel dev lint:watch test:watch start-mockapi",
+    "dev": "webpack-dev-server --config webpack.config.dev.babel.js --open",
+    "prod": "webpack-dev-server --config webpack.config.prod.babel.js --open",
     "open:src": "babel-node buildScripts/srcServer.js",
     "test": "jest",
     "test:watch": "npm run test -- --watch",
@@ -51,11 +53,10 @@
     "eslint-plugin-import": "^2.14.0",
     "eslint-watch": "^4.0.2",
     "express": "^4.16.3",
-    "extract-text-webpack-plugin": "^3.0.2",
     "file-loader": "^2.0.0",
-    "hard-source-webpack-plugin": "^0.12.0",
     "html-loader": "^0.5.5",
     "html-webpack-plugin": "^3.2.0",
+    "image-webpack-loader": "^4.3.1",
     "jest": "^23.6.0",
     "jest-cli": "^23.6.0",
     "jsdom": "^12.1.0",
@@ -75,8 +76,7 @@
     "style-loader": "^0.23.0",
     "webpack": "^4.20.2",
     "webpack-cli": "^3.1.2",
-    "webpack-dev-middleware": "^3.4.0",
-    "webpack-hot-middleware": "^2.24.2",
+    "webpack-dev-server": "^3.1.9",
     "webpack-md5-hash": "0.0.6",
     "xml-loader": "^1.2.1"
   }
