+++ src/index.html
@@ -29,5 +29,6 @@
 
     </tbody>
   </table>
+  <div class="robot"></div>
 </body>
 </html>
