--- webpack.config.dev.js
@@ -1,125 +1,0 @@
-import webpack from 'webpack';
-import HtmlWebpackPlugin from 'html-webpack-plugin';
-import path from 'path';
-import HardSourceWebpackPlugin from 'hard-source-webpack-plugin';
-//import CopyWebpackPlugin from 'copy-webpack-plugin';
-
-
-export default {
-  resolve: {
-    extensions: ['*', '.js', '.jsx', '.json']
-  },
-  devtool: 'cheap-module-eval-source-map', // more info:https://webpack.js.org/guides/development/#using-source-maps and https://webpack.js.org/configuration/devtool/
-  entry: [
-    path.resolve(__dirname, 'src/index.js') // Defining path seems necessary for this to work consistently on Windows machines.
-  ],
-  target: 'web',
-  mode: 'development',
-  output: {
-    path: path.resolve(__dirname, 'dist'), // Note: Physical files are only output by the production build task `npm run build`.
-    publicPath: '/',
-    filename: 'bundle.js'
-  },
-  plugins: [
-    new HardSourceWebpackPlugin(),
-    new webpack.HotModuleReplacementPlugin(),
-    new webpack.NoEmitOnErrorsPlugin(),
-    new HtmlWebpackPlugin({     // Create HTML file that includes references to bundled CSS and JS.
-      template: 'src/index.html',
-      minify: {
-        removeComments: true,
-        collapseWhitespace: true
-      },
-      inject: true
-    })
-  ],
-  module: {
-    rules: [
-      // transpile js/jsx files
-      {
-        test: /\.(js|jsx)?$/,
-        exclude: /node_modules/,
-        use: ['babel-loader']
-      },
-
-      // load fonts
-      {
-          test: /\.(woff|woff2|eot|ttf|otf)$/,
-          use: [
-              'file-loader'
-          ]
-      },
-
-      //
-      {
-        test: /\.svg(\?v=\d+\.\d+\.\d+)?$/,
-        use: [
-          {
-            loader: 'url-loader',
-            options: {
-              limit: 10000,
-              mimetype: 'image/svg+xml'
-            }
-          }
-        ]
-      },
-
-      // minify images
-      {
-        test: /\.(jpe?g|png|gif|ico)$/i,
-        use: [
-          {
-            loader: 'file-loader',
-            options: {
-              name: '[name].[ext]'
-            }
-          }
-        ]
-      },
-
-      // compile sass
-      {
-        test: /(\.css|\.scss|\.sass)$/,
-        use: [
-          'style-loader',
-          {
-            loader: 'css-loader',
-            options: {
-              sourceMap: true
-            }
-          }, {
-            loader: 'postcss-loader',
-            options: {
-              plugins: () => [
-                require('autoprefixer')
-              ],
-              sourceMap: true
-            }
-          }, {
-            loader: 'sass-loader',
-            options: {
-              includePaths: [path.resolve(__dirname, 'src', 'scss')],
-              sourceMap: true
-            }
-          }
-        ]
-      },
-
-      // load csv files
-      {
-        test: /\.(csv|tsv)$/,
-        use: [
-          'csv-loader'
-        ]
-      },
-
-      // load xml files
-      {
-        test: /\.xml$/,
-        use: [
-          'xml-loader'
-        ]
-      },
-    ]
-  }
-};
